+++ packages/preview-middleware-client/test/__mock__/sap/ui/core/ComponentContainer.ts
@@ -1,0 +1,7 @@
+import UIComponentMOck from './UIComponent';
+
+export default class {
+    getComponent() {
+        new UIComponentMOck();
+    }
+}